+++ packages/elements-angular/elements/ngcc.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ignorableDeepImportMatchers: [/@inovex.de\//]
+};