+++ packages/elements/src/components/ino-icon/icons.js
@@ -44,6 +44,7 @@ export default [
 	'phone',
 	'print',
 	'question',
+	'refresh',
 	'remove',
 	'report',
 	'search',