+++ packages/elements/src/components/ino-menu/MDCCustomMenu.js
@@ -6,9 +6,9 @@ export class MDCCustomMenu extends MDCMenu {
   // TODO: Test if basic onclick and keydown events still come trough
   constructor(mdcMenu) {
     super(mdcMenu);
-    this.foundation_.handleKeydown = (evt) => {};
-    this.foundation_.handleClick = (evt) => {};
-    this.menuSurface_.handleBodyClick_ = (evt) => {};
-    this.menuSurface_.handleKeydown_ = (evt) => {};
+    this.foundation.handleKeydown = (evt) => {};
+    this.foundation.handleClick = (evt) => {};
+    this.menuSurface_.handleBodyClick = (evt) => {};
+    this.menuSurface_.handleKeydown = (evt) => {};
   }
 }