+++ packages/elements/stencil.config.ts
@@ -3,8 +3,19 @@ import { sass } from '@stencil/sass';
 import { angularOutputTarget } from '@stencil/angular-output-target';
 
 export const config: Config = {
-  buildEs5: false, // see https://github.com/ionic-team/stencil/issues/253
+  buildEs5: 'prod',
+  extras: {
+    cssVarsShim: true,
+    dynamicImportShim: true,
+    shadowDomShim: true,
+    safari10: true,
+    scriptDataOpts: true,
+    appendChildSlotFix: false,
+    cloneNodeFix: false,
+    slotChildNodesFix: true,
+  },
   globalStyle: './src/components/styles/variables.scss',
+  globalScript: './src/util/import-fonts.ts',
   enableCache: true,
   namespace: 'inovex-elements',
   outputTargets: [